Create Credential passes status; Update Credential added
Create Credential now passes a status of 'issued' for the
self-authenticated credential, and of 'requested' for the
invited other authentication credential (which it also creates).

It stores the api_id it gets back on the ui authentication object;
yes, there's a difference between the API data model
(where what, in the UI data model is a credential having many authentications
is in the API a list of credentials, that is, the credential
links all three of the skill, the holder, and the authenticator),
and we haven't decided yet whether to keep the difference
(because from the point of view of the user, the UI data model
might be clearer) or to collapse it for consistency.

Right now, because only users have api_ids, it'll only create the
invited other authentication credential if that user already exists
in the system.  (We'll want to move api_id to contact or and have
credential creation also create an API user for contacts that aren't users yet,
or allow users which don't have accounts yet -- details tbd.)

We've also added an Update Credential API call such that when the user
accepts or rejects an authentication request it calls the API and
updates the credential status.