Email sufficient match for Contact.retrieve_or_build
Matching on name and email required the credentials_for_others add participants
to be using the same form for the name as the applicant did, else it
created a new contact, which meant participants didn't match if there was
an existing contact with that email but with a variant form for name.

Assume it should match.