Allow both \r\n and \n as line breaks in adding authenticators / participants
Oddly, it appears that on creating a credential, authenticators / participants
param textarea lines are separated by '\r\n', but in adding authenticators / participants
later in a different form, the param textarea lines are separated by '\n'.

This should probably be made consistent, but as an interim fix, accept either.
Also, because we were doing the split in multiple locations, pull the others
into NameAndEmail and only change it there.